Refactor header and improve performance
Now header is a custom element so it's reusable. But in the future I think it would be better to just have it as a part of the common page part and use Jekyll or even Astro (once I have my own page hosting).

Fonts are preconnected, and loaded later using JS, also header has a placeholder made out of img element (it can have settled the height which is a hack but it does the job, there's no layout shift). Also the script tag has been moved to the end of body